Cat

See merge request s05-webmobile2-sub3/S05P13D104!150

+++ backend/webcuration/src/main/java/com/web/webcuration/service/UserServiceImpl.java
@@ -8,6 +8,7 @@ import com.web.webcuration.dto.request.NickNameRequest;
 import com.web.webcuration.dto.request.ProfileRequest;
 import com.web.webcuration.dto.request.UserRequest;
 import com.web.webcuration.dto.response.BaseResponse;
+import com.web.webcuration.repository.UserQueryRepository;
 import com.web.webcuration.repository.UserRepository;
 import com.web.webcuration.utils.FileUtils;
 
@@ -23,6 +24,7 @@ public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
+    private final UserQueryRepository userQueryRepository;
 
     @Override
     public BaseResponse getUserInfo(String email) {
@@ -43,24 +45,29 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional
     public BaseResponse updateUser(ProfileRequest profileRequest) throws IllegalStateException, IOException {
-        Optional<User> user = userRepository.findById(profileRequest.getId());
-        if (user.isPresent()) {
-            User changeUser = user.get();
-            changeUser.setNickname(profileRequest.getNickname());
-            changeUser.setIntroduce(profileRequest.getIntroduce());
-            if (!changeUser.getImage().equals("/profileImg/user_image.png")) {
-                FileUtils.deleteProfile(changeUser.getImage());
+        if (userQueryRepository.DuplicateCheckName(profileRequest.getNickname())) {
+            Optional<User> user = userRepository.findById(profileRequest.getId());
+            if (user.isPresent()) {
+                User changeUser = user.get();
+                changeUser.setNickname(profileRequest.getNickname());
+                changeUser.setIntroduce(profileRequest.getIntroduce());
+                if (!changeUser.getImage().equals("/profileImg/user_image.png")) {
+                    FileUtils.deleteProfile(changeUser.getImage());
+                }
+                if (profileRequest.getImage() != null) {
+                    changeUser.setImage(FileUtils.uploadProfile(profileRequest.getImage()));
+                } else {
+                    // changeUser.setImage("C:\\Users\\Master\\Desktop\\img\\user_image.png");
+                    // 서버
+                    changeUser.setImage("/profileImg/user_image.png");
+                }
+                return BaseResponse.builder().status("200").status("success").data(userRepository.save(changeUser))
+                        .build();
             }
-            if (profileRequest.getImage() != null) {
-                changeUser.setImage(FileUtils.uploadProfile(profileRequest.getImage()));
-            } else {
-                // changeUser.setImage("C:\\Users\\Master\\Desktop\\img\\user_image.png");
-                // 서버
-                changeUser.setImage("/profileImg/user_image.png");
-            }
-            return BaseResponse.builder().status("200").status("success").data(userRepository.save(changeUser)).build();
+            throw new RuntimeException("수정하려는 유저가 없습니다.");
+        } else {
+            return BaseResponse.builder().status("500").msg("닉네임 중복").build();
         }
-        throw new RuntimeException("수정하려는 유저가 없습니다.");
     }
 
     @Override
@@ -78,9 +85,13 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public BaseResponse setNickname(NickNameRequest nicknameRequest) {
-        Optional<User> user = userRepository.findById(nicknameRequest.getId());
-        user.get().setNickname(nicknameRequest.getNickname());
-        return BaseResponse.builder().status("200").msg("success").data(userRepository.save(user.get())).build();
+        if (userQueryRepository.DuplicateCheckName(nicknameRequest.getNickname())) {
+            Optional<User> user = userRepository.findById(nicknameRequest.getId());
+            user.get().setNickname(nicknameRequest.getNickname());
+            return BaseResponse.builder().status("200").msg("success").data(userRepository.save(user.get())).build();
+        } else {
+            return BaseResponse.builder().status("500").msg("닉네임 중복").build();
+        }
     }
 
 }
